# docs/swagger.yaml
openapi: 3.0.3
info:
  title: Purchase Approval Platform API
  version: 1.0.0
  description: |
    REST API for creating purchase requests and completing multi-approver workflows.
    The deployed assessment environment exposes the API through AWS API Gateway and AWS Lambda.
servers:
  - url: https://jguzyqqsw9.execute-api.us-east-1.amazonaws.com/prod
    description: Deployed assessment environment
...
paths:
...
  /approvals/{token}/decision:
    post:
      tags: [Approvals]
      summary: Register approval decision
      operationId: processApprovalDecision
      parameters:
        - $ref: '#/components/parameters/Token'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DecisionRequest'
            examples:
              approved:
                value:
                  decision: APPROVED
              rejected:
                value:
                  decision: REJECTED
      responses:
        '200':
          description: Decision registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DecisionResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
components:
  parameters:
    Token:
      name: token
      in: path
      required: true
      description: Unique cryptographic approval token generated for the approver.
      schema:
        type: string
        minLength: 32
  responses:
    BadRequest:
      description: Invalid request or business rule violation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    DecisionRequest:
      type: object
      required: [decision]
      properties:
        decision:
          type: string
          enum: [APPROVED, REJECTED]
    DecisionResponse:
      type: object
      properties:
        approverId:
          type: string
          format: uuid
        requestId:
          type: string
          format: uuid
        status:
          type: string
          example: SIGNED
        signedAt:
          type: string
          format: date-time
        purchaseRequestStatus:
          type: string
          example: PENDING
    ErrorResponse:
      type: object
      properties:
        message:
          type: string